small redesign PopupTaskInfo && added asyncthunk getBoardThunk to boardPage
