+++ src/features/auth/login/model/slice/LoginSlice.ts
@@ -17,16 +17,13 @@ export const loginSlice = createSlice({
     },
     extraReducers: (builder) => {
         builder.addCase(fetchByIdStatus.pending, (state, action) => {
-            console.log(action.payload);
             state.isLoading = true;
         });
         builder.addCase(fetchByIdStatus.rejected, (state, action) => {
             state.isLoading = false;
-            console.log(action.payload);
             state.error = action.payload as string;
         });
-        builder.addCase(fetchByIdStatus.fulfilled, (state, action) => {
-            console.log(action.payload);
+        builder.addCase(fetchByIdStatus.fulfilled, (state) => {
             state.isLoading = false;
         });
     },
